Update rocket_codegen for latest nightly.

+++ codegen/src/utils/span_ext.rs
@@ -1,15 +1,15 @@
 use syntax::codemap::{Span, BytePos};
 
 pub trait SpanExt {
-    fn shorten_to(self, to_length: u32) -> Span;
+    fn shorten_to(self, to_length: usize) -> Span;
 
     /// Trim the span on the left and right by `length`.
     fn trim(self, length: u32) -> Span;
 }
 
 impl SpanExt for Span {
-    fn shorten_to(mut self, to_length: u32) -> Span {
-        self.hi = self.lo + BytePos(to_length);
+    fn shorten_to(mut self, to_length: usize) -> Span {
+        self.hi = self.lo + BytePos(to_length as u32);
         self
     }
